добавляет валидацию
