+++ index.html
@@ -241,7 +241,7 @@
                         class="text__hashtags"
                         type="text"
                         name="hashtags"
-                        placeholder="#ХэшТег"
+                        placeholder="#ХэшТег" required
                       />
                     </div>
                     <div class="img-upload__field-wrapper">
@@ -249,7 +249,7 @@
                         class="text__description"
                         name="description"
                         placeholder="Ваш комментарий..."
-                        maxlength="140"
+                        maxlength="140" required
                       ></textarea>
                     </div>
                   </fieldset>
@@ -455,6 +455,7 @@
         Загружаем...
       </div>
     </template>
+    <script src="pristine/pristine.min.js"></script>
     <script src="js/main.js" type="module"></script>
   </body>
 </html>
