+++ js/main.js
@@ -1,6 +1,6 @@
 import {addsPicturesOfUsers} from './preview.js';
 import './generator.js';
 import './full-screen.js';
-import './form-validation.js';
+import './validation-form.js';
 
 addsPicturesOfUsers();
